Move Symbol, TSSymbol, & SymbolType into TreeSitter.Symbol.
