+++ src/TreeSitter/Symbol.hs
@@ -1,8 +1,17 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 module TreeSitter.Symbol where
 
+import Data.Ix
 import Data.Word
-import TreeSitter.Language
 
-toSymbol :: forall symbol . Symbol symbol => Word16 -> symbol
+type TSSymbol = Word16
+
+toSymbol :: forall symbol . Symbol symbol => TSSymbol -> symbol
 toSymbol symbol = toEnum (min (fromIntegral symbol) (fromEnum (maxBound :: symbol)))
+
+
+data SymbolType = Regular | Anonymous | Auxiliary
+  deriving (Enum, Eq, Ord, Show)
+
+class (Bounded s, Enum s, Ix s, Ord s, Show s) => Symbol s where
+  symbolType :: s -> SymbolType
